requirements.txt
pdfplumber==0.11.4
python-docx==1.1.2
